feat: 完善 LeaveRoom 业务逻辑和业务 Webhook 文档
- 实现 LeaveRoom 四种情况的完整业务逻辑
  - 情况1: 发起者取消通话（MaxParticipants=2 且 creator=uid 且对方未加入）
  - 情况2: 参与者拒绝通话（MaxParticipants=2 且 status=0）
  - 情况3: 多人房间正常挂断
  - 情况4: 一对一房间正常挂断
- 新增房间状态 RoomStatusRejected (4)
- 统一参与者事件数据结构为 ParticipantEventData (RoomEventData + UID)
- 在 Swagger 文档中添加业务 Webhook 接口说明
  - 新增端点: POST /tgortc/webhook
  - 文档化所有房间事件和参与者事件类型
  - 添加完整的事件数据结构定义和示例

+++ internal/service/room_service.go
@@ -95,11 +95,11 @@ func (rs *RoomService) CreateRoom(req *models.CreateRoomRequest) (*models.Create
 		// 创建房间
 		room := models.Room{
 			SourceChannelID:   req.SourceChannelID,
-			SourceChannelType: int16(req.SourceChannelType),
+			SourceChannelType: req.SourceChannelType,
 			Creator:           req.Creator,
 			RoomID:            roomID,
-			RTCType:           int16(req.RTCType),
-			InviteOn:          int16(req.InviteOn),
+			RTCType:           req.RTCType,
+			InviteOn:          req.InviteOn,
 			Status:            models.RoomStatusNotStarted,
 			MaxParticipants:   maxParticipants,
 		}
